Add Option To Stage Server, Add Management Interface
The staging option allows all software to be downloaded to the server
prior to starting configuration, allowing configuration to be performed
without an Internet connection.

The management port provides an always available DHCPv4 port that can
allow local access for management if the other interfaces become
unresponsive. A mangemnent port is optional when the WAN port is DHCP
configured, and is mandatory when the WAN port is statically configured.